add CopyButton to shared components, move InvoiceModal to Transactions.
