+++ client/src/components/pages/Store/Wallet/Transactions/index.js
@@ -1,1 +1,2 @@
+export { InvoiceModal } from "./InvoiceModal";
 export { Transactions } from "./Transactions";
