+++ client/src/components/shared/CopyButton.jsx
@@ -1,0 +1,23 @@
+"use client";
+
+import { Button } from "@heroui/react";
+import { Copy } from "lucide-react";
+import { useTranslations } from "next-intl";
+
+import { copyToClipboard } from "@/components/pages/Store/Wallet/utils/formatters";
+
+export function CopyButton({ value, label, ...props }) {
+  const t = useTranslations("wallet");
+
+  return (
+    <Button
+      variant="bordered"
+      className="border border-border text-foreground hover:bg-muted disabled:opacity-50 disabled:cursor-not-allowed transition-colors"
+      onPress={() => copyToClipboard(value, t)}
+      {...props}
+    >
+      <Copy className="w-3 h-3 mr-1" />
+      {label}
+    </Button>
+  );
+}
